SAK-11544
Harmonize with SAK-12387 - this is almost like a tag of SAK-12387 
so that SAK-11544 can be checked out from message and mailarchive and
work with an otherwise clean trunk so I can get started having people
test SAK-11544.

After this SAK-12387 can move forward without affecting SAK-11544.  
This also means that any testing of SAK-11544 is also implicitly 
testing the first step in the SAK-12387 refactor.



git-svn-id: https://source.sakaiproject.org/svn/message/branches/SAK-11544@40567 66ffb92e-73f9-0310-93c1-f5514f145a0a